+++ tests/core/integration/auditlog/conftest.py
@@ -23,7 +23,7 @@ def auditlog_client():
         client = create_client()
         return client
     except Exception as e:
-        pytest.skip(f"AuditLog credentials not configured — skipping integration tests: {e}")
+        pytest.fail(f"Failed to create AuditLog client for cloud integration tests: {e}")
 
 
 @pytest.fixture
